DDM-637 - Making DBCSecrets module optional
Wanted to reduce the required number of dependencies. While DBCSecrets
is useful during development this dependency will no longer  be used
once the tests are deployed.  Instead secrets will come from env vars.
Have now configured things so that if all secrets will first be searched
for in environment variables, if the env vars are not defined the tests
will try to find the secrets.json file. If found will extract secrets
from there.  If not found then will raise a SecretsNotFound exception