Added PKG_TYPE env var to build
When a package is built, it runs the setup.py.  The setup.py calls the version.py that in turn hits pypi to find out what the last version of the package was. 

We are now building two packages though.  One called bcdc_apitests_dev which is built off the dev branch, and another called bcdc_apitests which builds off the master.  Versioning needs to know which package it is building, the dev or the master.  To resolve this for now creating an env var called PKG_TYPE.  That env var will be consumed by the version.py to determine which package to get the version from.  If the env var is not set the assumption is the package is dev.