DDM-910 Integration with DEV
During merge with DEV branch identified that the calculation of package
versions was always comming from the package that aligns with the name
of the module.

We are currently building two pypi packages: bcdc-apitests and
bcdc-apitests-dev.  When setup.py is run it needs to know which package
it is building.  Added an environment variable to the github actions
that builds the pypi package called PKG_TYPE.  If this is equal to
MASTER or PROD then the setup will build the "bcdc-apitests" package.
If its equal to DEV then it builds the 'bcdc-apitests-dev package